+++ sdk/identity/identity/test/internal/node/azurePowerShellCredential.spec.ts
@@ -11,7 +11,7 @@ import {
 import { AzurePowerShellCredential } from "../../../src";
 import { GetTokenOptions } from "@azure/core-auth";
 import Sinon from "sinon";
-import { assert } from "chai";
+import { assert } from "@azure/test-utils";
 import { commandStack } from "../../../src/credentials/azurePowerShellCredential";
 import { processUtils } from "../../../src/util/processUtils";
 
@@ -26,7 +26,8 @@ function resetCommandStack(): void {
 
 describe("AzurePowerShellCredential", function () {
   const scope = "https://vault.azure.net/.default";
-
+  const tenantIdErrorMessage =
+    "Invalid tenant id provided. You can locate your tenant id by following the instructions listed here: https://docs.microsoft.com/partner-center/find-ids-and-domain-names.";
   afterEach(() => {
     resetCommandStack();
   });
@@ -230,4 +231,43 @@ describe("AzurePowerShellCredential", function () {
     assert.ok(token?.token);
     assert.ok(token?.expiresOnTimestamp!);
   });
+
+  for (const tenantId of [
+    "&quot;invalid-tenant-id&quot;",
+    " ",
+    "12345678-1234-1234-1234-123456789012|",
+    "12345678-1234-1234-1234-123456789012 |",
+    "<",
+    ">",
+    "\0",
+    "<12345678-1234-1234-1234-123456789012>",
+    "12345678-1234-1234-1234-123456789012&",
+    "12345678-1234-1234-1234-123456789012;",
+    "12345678-1234-1234-1234-123456789012,",
+  ]) {
+    it(`rejects invalid tenant id "${tenantId}" in getToken`, async function () {
+      const credential = new AzurePowerShellCredential();
+      await assert.isRejected(
+        credential.getToken("https://service/.default", {
+          tenantId: tenantId,
+        }),
+        tenantIdErrorMessage
+      );
+    });
+    it(`rejects invalid tenant id "${tenantId}" in constructor`, function () {
+      assert.throws(() => {
+        new AzurePowerShellCredential({ tenantId: tenantId });
+      }, tenantIdErrorMessage);
+    });
+  }
+
+  for (const inputScope of ["scope |", "", "\0", "scope;", "scope,", "scope'", "scope&"]) {
+    it(`rejects invalid scope "${inputScope}"`, async function () {
+      const credential = new AzurePowerShellCredential();
+      await assert.isRejected(
+        credential.getToken(inputScope),
+        "Invalid scope was specified by the user or calling client"
+      );
+    });
+  }
 });